+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "concurrently \"yarn run server\" \"yarn run client\"",
+    "dev": "concurrently \"yarn server\" \"yarn client\"",
     "client": "webpack-dev-server --config webpack.client.dev.js",
     "eslint-fix": "eslint --fix \"src/**/*.js\" && eslint --fix \"server/**/*.js\"",
     "format": "prettier --write \"src/**/*.js\"",
@@ -24,7 +24,7 @@
   "dependencies": {
     "@material-ui/core": "^4.4.3",
     "concurrently": "^4.1.2",
-    "dotenv": "^8.1.0",
+    "dotenv-extended": "^2.4.0",
     "express": "^4.17.1",
     "history": "^4.10.1",
     "html-webpack-plugin": "^3.2.0",
