+++ server/models/User.js
@@ -1,18 +1,24 @@
 import mongoose from 'mongoose';
 
-const UserSchema = new mongoose.Schema({
-  email: {
-    type: String,
-    required: true,
+const UserSchema = new mongoose.Schema(
+  {
+    email: {
+      type: String,
+      required: true,
+      unique: true,
+    },
+    passwordHash: {
+      type: String,
+      required: true,
+    },
+    passwordSalt: {
+      type: String,
+      required: true,
+    },
   },
-  passwordHash: {
-    type: String,
-    required: true,
-  },
-  passwordSalt: {
-    type: String,
-    required: true,
-  },
-});
+  {
+    timestamps: true,
+  }
+);
 
 export default mongoose.model('User', UserSchema);
